+++ ci/run_test_post_always.sh
@@ -6,24 +6,27 @@
 set -uex
 
 # shellcheck disable=SC1091
-# shellcheck disable=SC2029
 source ./.build_vars.sh
 DAOS_BASE="${SL_PREFIX%/install*}"
 NODE="${NODELIST%%,*}"
 mydir="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)"
 
-if [ "$1" == "memcheck" ]; then
-    echo "run test post always with memcheck"
-    rm -rf run_test_memcheck.sh
-    WITH_VALGRIND=memcheck
-    ssh "$SSH_KEY_ARGS" jenkins@"$NODE" \
-      "DAOS_BASE=$DAOS_BASE             \
-      WITH_VALGRIND=$WITH_VALGRIND      \
-      $(cat "$mydir/run_test_post_memcheck_node.sh")"
+if [ $# -ge 1 ] && [ -n "$1" ]; then
+    if [ "$1" == "memcheck" ]; then
+        echo "run test post always with memcheck"
+        rm -rf run_test_memcheck.sh
+        WITH_VALGRIND=memcheck
+        # shellcheck disable=SC2029
+        ssh "$SSH_KEY_ARGS" jenkins@"$NODE" \
+          "DAOS_BASE=$DAOS_BASE             \
+          WITH_VALGRIND=$WITH_VALGRIND      \
+          $(cat "$mydir/run_test_post_memcheck_node.sh")"
+    fi
 else
     echo "run test post always"
     rm -rf run_test.sh vm_test
 
+    # shellcheck disable=SC2029
     ssh "$SSH_KEY_ARGS" jenkins@"$NODE" \
       "DAOS_BASE=$DAOS_BASE             \
       $(cat "$mydir/run_test_post_always_node.sh")"
